+++ src/model/board.ts
@@ -2,9 +2,12 @@ import { FetchError, GET } from '@/utils/fetch'
 import { Success, Try } from '@/utils/fp/Try'
 import { IBaseBoard, IBoard, ITag } from '@cc98/api'
 import { Container } from '@cc98/state'
-
+interface T {
+  name: string
+  id: number
+}
 interface State {
-  tagData: ITag[],
+  tagData: T[],
   boardData: IBaseBoard[]
 }
 
@@ -17,6 +20,7 @@ export class BoardInfoStore extends Container<State> {
   constructor() {
     super()
     this.getInfo()
+    this.getTagInfo()
   }
   /**
    * 获取版面信息
@@ -30,7 +34,18 @@ export class BoardInfoStore extends Container<State> {
 
     return this.forceGetInfo()
   }
+  /**
+   * 获取标签信息
+   * @return {Promise<Try<IUser, FetchError>>}
+   * @memberof BoardInfoStore
+   */
+  getTagInfo = (): Promise<Try<T[], FetchError>> => {
+    if (this.state.boardData.length !== 0) {
+      return Promise.resolve(Try.of<T[], FetchError>(Success.of(this.state.tagData)))
+    }
 
+    return this.forceGetTagInfo()
+  }
   /**
    * 强制从服务器获取全部信息
    * @return {Promise<Try<IUser, FetchError>>}
@@ -43,9 +58,23 @@ export class BoardInfoStore extends Container<State> {
 
     return res
   }
+  /**
+   * 强制从服务器获取全部信息
+   * @return {Promise<Try<IUser, FetchError>>}
+   * @memberof BoardInfoStore
+   */
+  forceGetTagInfo = async (): Promise<Try<T[], FetchError>> => {
+    const res = await GET<T[]>('config/global/alltag')
 
+    res.fail().succeed(info => this.setTagInfo(info))
+
+    return res
+  }
   setInfo = (data: IBaseBoard[]) => {
     this.put(state => state.boardData = data)
+  }
+  setTagInfo = (data: T[]) => {
+    this.put(state => state.tagData = data)
   }
 }
